dj - Corrected CourseShowPage.stories.js to display individual course, and to follow naming convention
